Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    float n1, n2;
+    int n1, n2;
     char op;
-    scanf("%f %f %c", &n1, &n2, &op);
+    scanf("%d %d %c", &n1, &n2, &op);
     if (op == '+'){
         r = n1 + n2;
     }
